+++ src/main/java/com/moa/backend/domain/wallet/entity/MakerWallet.java
@@ -1,6 +1,6 @@
 package com.moa.backend.domain.wallet.entity;
 
-import com.moa.backend.domain.user.entity.Maker;
+import com.moa.backend.domain.maker.entity.Maker;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -12,10 +12,11 @@ import jakarta.persistence.OneToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
-import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
